Update base.config
Without adding the chunk of code at the top, a user trying to set up a custom config by using the -c flag will hit a hard wall when they try to change anything. Nf-flow will immediately crash. This should be seen as a temporary fix until a more elegant solution can be thought of.